Protect react routes

+++ frontend/src/components/WorkoutDetails.js
@@ -5,14 +5,24 @@ import { useNavigate } from 'react-router-dom'
 import { FaEdit} from 'react-icons/fa'
 import { FaTrashAlt } from 'react-icons/fa'
 import  FormatDistancetToNow  from 'date-fns/formatDistanceToNow'
+import { useAuthContext } from '../hooks/useAuthContext'
 
 const WorkoutDetails = ({ workout }) => {
     const { dispatch } = useWorkoutsContext();
+    const { user } = useAuthContext();
     const navigate = useNavigate();
 
     const handleDelete = async () => {
+
+        if(!user) {
+            return
+        }
+
         const response = await fetch(`/api/workouts/${workout._id}`, {
-        method: 'DELETE'
+        method: 'DELETE',
+        headers: {
+            'Authorization': `Bearer ${user.token}`
+        }
         })
         const json = await response.json()
